+++ frontend/src/components/Worksheet/URLExportButton.tsx
@@ -1,0 +1,36 @@
+import { compressToEncodedURIComponent } from 'lz-string';
+import { toast } from 'react-toastify';
+import {
+  useWorksheet,
+  type ExoticWorksheet,
+} from '../../contexts/worksheetContext';
+
+export default function URLExportButton() {
+  const { viewedSeason, viewedWorksheetNumber, courses } = useWorksheet();
+
+  async function handleExport() {
+    const payload: ExoticWorksheet = {
+      season: viewedSeason,
+      name:
+        viewedWorksheetNumber === 0
+          ? 'Main Worksheet'
+          : `Worksheet ${viewedWorksheetNumber}`,
+      courses: courses.map((c) => ({
+        crn: c.listing.crn,
+        hidden: c.hidden ?? false,
+        color: c.color,
+      })),
+    };
+
+    await navigator.clipboard.writeText(
+      `${window.location.origin}/worksheet?ws=${compressToEncodedURIComponent(JSON.stringify(payload))}`,
+    );
+    toast.success('Worksheet copied to clipboard as URL');
+  }
+
+  return (
+    <button type="button" onClick={handleExport}>
+      Copy worksheet as URL
+    </button>
+  );
+}
